refactor: Kafka 역할별 어댑터 패키지 분리
