+++ nolzo-api/src/test/java/com/noljo/nolzo/notification/adapter/producer/KafkaSeatAvailableEventPublisherTest.java
@@ -1,7 +1,8 @@
-package com.noljo.nolzo.notification.adapter.kafka;
+package com.noljo.nolzo.notification.adapter.producer;
 
 import static org.mockito.Mockito.verify;
 
+import com.noljo.nolzo.notification.adapter.message.SeatAvailableEventMessage;
 import com.noljo.nolzo.notification.domain.event.SeatAvailableEvent;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.Test;
