+++ nolzo-infra/src/main/java/com/noljo/nolzo/notification/adapter/message/NotificationBatchRequestMessage.java
@@ -1,4 +1,4 @@
-package com.noljo.nolzo.notification.adapter.kafka;
+package com.noljo.nolzo.notification.adapter.message;
 
 import com.noljo.nolzo.notification.domain.event.NotificationBatchRequest;
 import java.time.LocalDateTime;
